+++ .template-lintrc.js
@@ -10,6 +10,7 @@ module.exports = {
         'no-bare-strings': true,
         'no-implicit-this': true,
         'no-nested-interactive': false,
+        quotes: 'single',
     },
 
     ignore: [