+++ app/adapters/user-setting.ts
@@ -1,0 +1,11 @@
+import OsfAdapter from './osf-adapter';
+
+export default class UserSettingAdapter extends OsfAdapter {
+    parentRelationship = 'user';
+}
+
+declare module 'ember-data' {
+    interface AdapterRegistry {
+        'user-setting': UserSettingAdapter;
+    }
+}